move back to `bevy_ptr`
don't need a specific impl for World apparently
